Revert n conversion to double in r35422 for mitab_geometry.cpp.
Accidentally converted n from an int to a double when breaking out all
the local vars.

mitab_geometry.cpp:169:43: error: implicit conversion turns floating-point
      number into integer: 'double' to 'size_t' (aka 'unsigned long')
      [-Werror,-Wfloat-conversion]
    double *xintersect = (double *)calloc(n, sizeof(double));


git-svn-id: https://svn.osgeo.org/gdal/trunk@35424 f0d54148-0727-0410-94bb-9a71ac55c965

+++ gdal/ogr/ogrsf_frmts/mitab/mitab_geometry.cpp
@@ -157,7 +157,7 @@ int OGRPolygonLabelPoint(OGRPolygon *poPoly, OGRPoint *poLabelPoint)
 
     double skip = (oEnv.MaxY - oEnv.MinY)/NUM_SCANLINES;
 
-    double n = 0;
+    int n = 0;
     for( int j = 0; j < OGR_NUM_RINGS(poPoly); j++ )
     {
         /* count total number of points */
